renames distcache -> remotecache
