+++ conf/defaults.ini
@@ -107,7 +107,7 @@ path = grafana.db
 cache_mode = private
 
 #################################### Cache server #############################
-[cache_server]
+[remote_cache]
 # Either "redis", "memcached" or "database" default is "database"
 type = database
