+++ pkg/infra/remotecache/database_storage_test.go
@@ -1,4 +1,4 @@
-package distcache
+package remotecache
 
 import (
 	"testing"
