+++ pkg/setting/setting.go
@@ -242,7 +242,7 @@ type Cfg struct {
 	EditorsCanOwn bool
 
 	// DistributedCache
-	CacheOptions *CacheOpts
+	RemoteCacheOptions *RemoteCacheOptions
 }
 
 type CommandLineArgs struct {
@@ -782,8 +782,8 @@ func (cfg *Cfg) Load(args *CommandLineArgs) error {
 	enterprise := iniFile.Section("enterprise")
 	cfg.EnterpriseLicensePath = enterprise.Key("license_path").MustString(filepath.Join(cfg.DataPath, "license.jwt"))
 
-	cacheServer := iniFile.Section("cache_server")
-	cfg.CacheOptions = &CacheOpts{
+	cacheServer := iniFile.Section("remote_cache")
+	cfg.RemoteCacheOptions = &RemoteCacheOptions{
 		Name:    cacheServer.Key("type").MustString("database"),
 		ConnStr: cacheServer.Key("connstr").MustString(""),
 	}
@@ -791,7 +791,7 @@ func (cfg *Cfg) Load(args *CommandLineArgs) error {
 	return nil
 }
 
-type CacheOpts struct {
+type RemoteCacheOptions struct {
 	Name    string
 	ConnStr string
 }
